backend/requirements.txt
opencv-python-headless==4.9.0.80
numpy==1.26.4
pydantic==2.6.4
opencv-python-headless==4.9.0.80
numpy==1.26.4
pydantic==2.6.4
networkx==3.2.1
scipy==1.12.0
google-generativeai==0.8.3
python-dotenv==1.0.1